fix php code and comment

+++ app/Http/Controllers/TeacherController.php
@@ -7,6 +7,7 @@ use App\Group;
 use App\OtherHoure;
 use App\TableHoure;
 use App\Teacher;
+use App\Course;
 use http\Env\Response;
 use Illuminate\Http\Request;
 
@@ -25,60 +26,113 @@ class TeacherController extends Controller
         ]);
     }
 
-    public function show($id)
-    {
-        // получаем преподавателя
-        $teacher = Teacher::find($id);
+    /**
+     * идентификатор преподаваетля
+     * @param $id
+     *
+     *  Вернет предметы преподавателя и количество часов каждого предмета и общее
+     * @return array
+     */
 
-        // получаем все групы
-        $groups = Group::get();
+    private function generate_date_show($id)
+    {
+        $teacher = [];
+
+        // Проверяем наличие преподователя по ID
+        if (Teacher::find($id) == null) {
+            return null;
+        }
+
+        // получаем преподавателя
+        $teacher['teacher_info'] = Teacher::find($id);
+
+        // если не нашёл часов у преподавателя то NULL
+        if (!Teacher::find($id)->hours()->exists()) {
+            return null;
+        }
 
         // получаем все часы преподавателя
         $teacher_hour = Teacher::find($id)->hours()->get();
 
-        // Получаем дисциплины
-        $disciplines = Discipline::get();
-
         // групируем по предметам
         $collection_items = $teacher_hour->groupBy('discipline_id');
 
-        // получаем тарификацию преподавателя
-        $tarificatin = Teacher::find($id)->other_hour()->first();
+        // если не нашёл дополнительных часов у преподавателя то NULL
+        if (!Teacher::find($id)->other_hour()->exists()) {
+            return null;
+        }
 
-        // часы за каждый предмет
-        $hour = [];
+        // получаем тарификацию преподавателя
+        $teacher['other_hour'] = Teacher::find($id)->other_hour()->first();
 
         // количество предметов
-        $count_discipline = count($collection_items);
+        $teacher['count_discipline'] = count($collection_items);
 
         // общие количество часов c группами
-        $sum_hour_group = 0;
+        $teacher['sum_hour_group'] = 0;
 
         if (count($collection_items) > 0) {
             foreach ($collection_items as $i => $item) {
-                $hour[$i]['discipline'] = Discipline::find($collection_items[$i][0]['discipline_id'])->name;
-                $hour[$i]['discipline_id'] = Discipline::find($collection_items[$i][0]['discipline_id'])->id;
+                $teacher['discipline'][$i]['discipline'] = Discipline::find($collection_items[$i][0]['discipline_id'])->name;
+                $teacher['discipline'][$i]['discipline_id'] = Discipline::find($collection_items[$i][0]['discipline_id'])->id;
                 foreach ($item as $j => $node) {
-                    $hour[$i]['time'][$j]['hour'] = $item[$j]['hour'];
-                    $hour[$i]['time'][$j]['group'] = Group::find($item[$j]['group_id'])->name;
-                    $hour[$i]['time'][$j]['id'] = $node->id;
+                    $teacher['discipline'][$i]['time'][$j]['hour'] = $item[$j]['hour'];
+                    $teacher['discipline'][$i]['time'][$j]['group'] = Group::find($item[$j]['group_id'])->name;
+                    $teacher['discipline'][$i]['time'][$j]['id'] = $node->id;
                 }
-                $sum_hour_group += $hour[$i]['sum_hour'] = $collection_items[$i]->sum('hour');
+                $teacher['sum_hour_group'] += $teacher['discipline'][$i]['sum_hour'] = $collection_items[$i]->sum('hour');
             }
         }
 
         // получаем все часы
-        $total = $sum_hour_group + array_sum($tarificatin->toArray());
+        $teacher['total_hour'] = $teacher['sum_hour_group'] + array_sum($teacher['other_hour']->toArray());
+
+        return $teacher;
+    }
+
+
+    public function show($id)
+    {
+        // получаем преподавателя и его часы
+        $teacher = $this->generate_date_show($id);
+
+        if ($teacher == null) {
+            return redirect()->route('entry.home')
+                ->withErrors([
+                    'error' => 'Преподаватель не найден'
+                ]);
+        }
+
+        // получаем все групы
+        $groups = Group::get();
+
+        // получаем все дисциплины
+        $disciplines = Discipline::get();
+
+        // групируем по курсам
+        $course_group = $groups->groupBy('course_id');
+
+        // список курсов и их групп
+        $group_list = [];
+
+        if (count($course_group) > 0) {
+            foreach ($course_group as $i => $item) {
+                $group_list[$i]['course_name'] = Course::find($i)->name;
+                $group = $item->toArray();
+                $group_list[$i]['groups'] = $group;
+                $group_list[$i]['count_grups'] = count($group);
+            }
+        }
 
         return view('teacher.show', [
-            'teacher' => $teacher,
-            'groups' => $groups,
+            'teacher' => $teacher['teacher_info'],
+            'group_list' => $group_list,
             'disciplines' => $disciplines,
-            'hour' => $hour,
-            'other_hour' => $tarificatin,
-            'sum_hour_group' =>$sum_hour_group,
-            'count_discipline' => $count_discipline,
-            'total' => $total
+            'hour' => $teacher['discipline'],
+            'other_hour' => $teacher['other_hour'],
+            'sum_hour_group' => $teacher['sum_hour_group'],
+            'count_discipline' => $teacher['count_discipline'],
+            'total' => $teacher['total_hour']
         ]);
     }
 
@@ -87,8 +141,7 @@ class TeacherController extends Controller
         $value = $request->value;
         $teacher_id = $request->teacher_id;
 
-        if (empty($value))
-        {
+        if (empty($value)) {
             $value = 0;
         }
 
@@ -111,9 +164,41 @@ class TeacherController extends Controller
         ], 200);
     }
 
+    // Генерация EXEL данных
+    private function generate_exel_table($data)
+    {
+        echo "<table class=\"table table-bordered bg-white table-striped\">";
+        echo "<thead>";
+        echo "<tr>";
+        echo "<td rowspan=\"2\">Дисциплина</td>";
+
+        echo "</tr>";
+        echo "</table>";
+    }
+
+    // вывод в excel файл
+    public function print_excel($id)
+    {
+        $data = $this->generate_date_show($id);
+       
+        header ( "Content-type: application/vnd.ms-excel" );
+        header ( "Content-Disposition: attachment; filename=foo_bar.xls" );
+
+        $this->generate_exel_table($data);
+    }
+
+
     public function edit($id)
     {
         $teacher = Teacher::find($id);
+
+        if ($teacher == null) {
+            return redirect()->route('entry.home')
+                ->withErrors([
+                    'error' => 'Преподаватель не найден'
+                ]);
+        }
+
         return view('layouts.edit', [
             'teacher' => $teacher
         ]);
@@ -122,6 +207,14 @@ class TeacherController extends Controller
     public function delete($id)
     {
         $teacher = Teacher::find($id);
+
+        if ($teacher == null) {
+            return redirect()->route('entry.home')
+                ->withErrors([
+                    'error' => 'Преподаватель не найден'
+                ]);
+        }
+
         $teacher->delete();
         return redirect()->route('teacher.home');
     }
@@ -161,8 +254,7 @@ class TeacherController extends Controller
         $entry_id = $request->entry_id;
         $value = $request->value;
 
-        if (empty($value))
-        {
+        if (empty($value)) {
             $value = 0;
         }
 
@@ -175,7 +267,7 @@ class TeacherController extends Controller
         $discipline_id = $entry->discipline_id;
 
         // количество часов по предмету
-        $sum = TableHoure::where('teacher_id', $teacher_id)->where('discipline_id',$discipline_id)->sum('hour');
+        $sum = TableHoure::where('teacher_id', $teacher_id)->where('discipline_id', $discipline_id)->sum('hour');
 
         // количество часов по предметам
         $sum_hour_group = TableHoure::where('teacher_id', $teacher_id)->sum('hour');
@@ -186,7 +278,7 @@ class TeacherController extends Controller
         $total = $sum_hour_group + array_sum($tarificatin->toArray());
 
         return response([
-            'sum'=> $sum,
+            'sum' => $sum,
             'sum_hour_group' => $sum_hour_group,
             'total' => $total
         ], 200);
